+++ cmd/neofs-node/config/policer/config_test.go
@@ -15,12 +15,18 @@ func TestPolicerSection(t *testing.T) {
 		empty := configtest.EmptyConfig()
 
 		require.Equal(t, policerconfig.HeadTimeoutDefault, policerconfig.HeadTimeout(empty))
+		require.Equal(t, policerconfig.ReplicationCooldownDefault, policerconfig.ReplicationCooldown(empty))
+		require.Equal(t, uint32(policerconfig.ObjectBatchSizeDefault), policerconfig.ObjectBatchSize(empty))
+		require.Equal(t, uint32(policerconfig.MaxWorkersDefault), policerconfig.MaxWorkers(empty))
 	})
 
 	const path = "../../../../config/example/node"
 
 	var fileConfigTest = func(c *config.Config) {
 		require.Equal(t, 15*time.Second, policerconfig.HeadTimeout(c))
+		require.Equal(t, 101*time.Millisecond, policerconfig.ReplicationCooldown(c))
+		require.Equal(t, uint32(11), policerconfig.ObjectBatchSize(c))
+		require.Equal(t, uint32(21), policerconfig.MaxWorkers(c))
 	}
 
 	configtest.ForEachFileType(path, fileConfigTest)