+++ pkg/services/policer/process.go
@@ -20,11 +20,23 @@ func (p *Policer) Run(ctx context.Context) {
 }
 
 func (p *Policer) shardPolicyWorker(ctx context.Context) {
+	p.cfg.RLock()
+	repCooldown := p.repCooldown
+	batchSize := p.batchSize
+	p.cfg.RUnlock()
+
 	var (
 		addrs  []objectcore.AddressWithType
 		cursor *engine.Cursor
 		err    error
 	)
+
+	t := time.NewTimer(repCooldown)
+	defer func() {
+		if !t.Stop() {
+			<-t.C
+		}
+	}()
 
 	for {
 		select {
@@ -33,7 +45,7 @@ func (p *Policer) shardPolicyWorker(ctx context.Context) {
 		default:
 		}
 
-		addrs, cursor, err = p.jobQueue.Select(cursor, p.batchSize)
+		addrs, cursor, err = p.jobQueue.Select(cursor, batchSize)
 		if err != nil {
 			if errors.Is(err, engine.ErrEndOfListing) {
 				time.Sleep(time.Second) // finished whole cycle, sleep a bit
@@ -55,16 +67,10 @@ func (p *Policer) shardPolicyWorker(ctx context.Context) {
 				}
 
 				err = p.taskPool.Submit(func() {
-					lastTime, ok := p.cache.Get(addr.Address)
-					if ok && time.Since(lastTime) < p.evictDuration {
-						return
-					}
-
 					p.objsInWork.add(addr.Address)
 
 					p.processObject(ctx, addr)
 
-					p.cache.Add(addr.Address, time.Now())
 					p.objsInWork.remove(addr.Address)
 				})
 				if err != nil {
@@ -72,10 +78,23 @@ func (p *Policer) shardPolicyWorker(ctx context.Context) {
 				}
 			}
 		}
+
+		select {
+		case <-ctx.Done():
+			return
+		case <-t.C:
+			p.cfg.RLock()
+			t.Reset(p.repCooldown)
+			p.cfg.RUnlock()
+		}
 	}
 }
 
 func (p *Policer) poolCapacityWorker(ctx context.Context) {
+	p.cfg.RLock()
+	maxCapacity := p.maxCapacity
+	p.cfg.RUnlock()
+
 	ticker := time.NewTicker(p.rebalanceFreq)
 	for {
 		select {
@@ -84,7 +103,7 @@ func (p *Policer) poolCapacityWorker(ctx context.Context) {
 			return
 		case <-ticker.C:
 			neofsSysLoad := p.loader.ObjectServiceLoad()
-			newCapacity := int((1.0 - neofsSysLoad) * float64(p.maxCapacity))
+			newCapacity := int((1.0 - neofsSysLoad) * float64(maxCapacity))
 			if newCapacity == 0 {
 				newCapacity++
 			}